+++ server/auth/facebook/passport.js
@@ -27,7 +27,7 @@ export function setup(User, config) {
           provider: 'facebook',
           facebook: profile._json
         });
-        user.save()
+        return user.save()
           .then(savedUser => done(null, savedUser))
           .catch(err => done(err));
       })